\name{sscsample.data}
\alias{sscsample.data}
\title{A stratified and clustered data set}
\description{A data set with predetermined strata and clusters. This data set is mostly for internal use with the sscsample function}
\usage{data(sscsample.data)}
\format{
	\describe{
		\item{value:}{the measurement taken on each sample unit}
		\item{stratum:}{the stratum the sample unit belongs to}
		\item{cluster:}{the cluster the sample unit belongs to}
	}
}
\keyword{datasets}

